fix:!firefox install hypr-key-repeat
fix(env):firefox moz wayland 1

fix(rofi) wallpapaer select ignore empty

fix(install) nvim extra

fix install wrong pkg name

fix instlall keyd failed

fix git config core editor nvim

fix hypr key repat delay 200

fix hypr key repeat rate 45

Revert "fix(env):firefox moz wayland 1"

This reverts commit a7194d613e5755ebf8f6bb33eb1124008d05abee.

+++ Install/KeydInstall.sh
@@ -1,11 +1,17 @@
 read -p "Install laptop conf?" ilap
-yay -S --noconfirm  --needed gcc
+yay -S --noconfirm --needed gcc
 mkdir -p ~/Temp
 git clone https://github.com/rvaiya/keyd ~/Temp/keyd
+
+if [[ ! $? -e 0 ]]; then
+	echo "clone failed"
+	exit 1
+fi
+
 cd ~/Temp/keyd
 make && sudo make install
 if [[ $ilap = [Yy] ]]; then
-  # TODO: this path is pretty bad
+	# TODO: this path is pretty bad
 	sudo ln -sf ~/dotfiles/keyd/laptop.conf /etc/keyd/laptop.conf
 fi
 sudo ln -sf ~/dotfiles/keyd/redox.conf /etc/keyd/redox.conf
